#3764 comment: split test_appointments.py into several files
